requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
asttokens==3.0.0
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
click==8.3.1
colabcode==0.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
discord.py==2.6.4
executing==2.2.0
Flask==3.1.2
fonttools==4.56.0
frozenlist==1.8.0
idna==3.11
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.7.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pefile==2024.8.26
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.4.1
pure_eval==0.2.3
pygame==2.6.1
pygame-ce==2.5.3
Pygments==2.19.1
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyngrok==7.2.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
tzdata==2025.1
wcwidth==0.2.13
Werkzeug==3.1.5
yarl==1.22.0
